Initial protocol implementation with create log command
Adds the initial implementation for the rog protocol, because the
protocol is built to read binary the main program was separated in two
different binaries a server and a cli, the cli has the command to create
a log and the server implements log creation. Initially only 4kb
messages will be processed.