Split log in multiple files
One of the problems with the past implementation was that all messages
were stored in a single file, because there's only one writer that's not
such a big problem, but it ends up being very hard parse the file the
bigger it gets. With this implementation rog tries to keep the log files
at around 8kB, messages are stored on the file by their id. This way in
the next step the messages will be stored in a different format, a
format that allows the readers to not load the entire file in memory.