Log file format reestructure
Creates a new file format for the log files. This new format allows rog
to fetch messages without loading the whole file in memory, this way the
read are more efficient. The writes are also more efficient because the
old format was just a HashMap with the messages read serialized with
bincode, instead load loading the whole file in memory, serializing the
HashMap and all the messages in that file, the write process now only needs
to append the record in the file.