Protocol revision
Reestructure the protocol to be more "binary friendly", the old protocol
was built with RESP in my mind, given that I built something similar on
ssache(eaneto/ssache). But RESP works better when dealing with strings,
and rog messages are sent as bytes, so that the clients can send files
without the need to transform them into text. The new protocol is a
little more restrict in some ways, but much easier to parse and build
clients. I started documenting the protocol in the README, at least for
the requests, the responses are still undocumented.