Response revision and standardization (#18)
Rewrite all responses to follow the same structure. The first byte of
any response indicates if there were any errors processing the request,
if the byte is 0 there were no errors and if it is one there was some
kind of error. In the case of the create log and publish commands the
successful response only returns one byte, on the fetch command the
first byte is 0, the following 8 bytes the message size and the
following N bytes the actual messages fetched from the log.

All commands have the same error response structure, with the first byte
being 1, the next 8 bytes the message size and the following N bytes the
actual error message.