Keep log filenames in-memory (#22)
Before this change, every time a new message was fetched or written to
disk all log files in the partition directory were listed to pick the
log file for the message id. Now all filenames are kept in memory, they
are loaded with the logs when the server is starts, when a new log file
is created the id is added to the structure.