Implements ack command and remove offset incrementation from fetch
Separates the offset incrementation from the fetch command, this is done
so that the if the client is processing the fetched and message and for
some reason encounters an error it'll be possible to consume the same
message again. This way the clients have to make a new request to only
acknowledge the message, incrementing the log group offset.